refactor: replace django-address with cities_light
django-address does not use abstract base classes for models so it was
impossible to add a UniqueConstraint and meet one of the requirements
for duplicate addresses.

I tried to use a reasonably flexible Address model within the
application but pull in cities_light for city, region and country data
which also makes it easy to prepopulate the API with some data for those
models.

City data limited to the UK to save some import time for this example.