feat: reject duplicate addresses
Use Django's UniqueConstraint on all address fields, as well as a DRF
serializer validator to provide a better 400 response to the client.

This is pretty loose validation and is still likely to lead to a
multitude of duplicate addresses, but this is still probably a better
user experience than stricter uniqueness checks.