chore: dockerise application
Ran into a strange migration issue where Django insists on creating a
new migration for the cities_light dependency after it's installed. This
was a requirement for the first migration in this project, and it was
validating the existence of the previous migration before creating it.

Removing the migration dependency (it's not clear if there is one in
reality) solves it, alongside running cities_light makemigrations and
migrate separately.