feat: istio jwt authentication
As per the Istio docs at
https://istio.io/latest/docs/tasks/security/authentication/authn-policy/#end-user-authentication
this sets up some networking and request authentication based on JWTs,
using a root of trust JWKS and client JWT the Istio team made available
publicly for testing.

The idea is to allow Istio to handle authentication transparently and in
a uniform fashion across multiple services, passing user context along
to the service (after checking the JWT) to aid database filtering e.g.
addresses for a user (see the outputPayloadToHeader option in
requestauth.yaml in the JWT Rule, defined:
https://istio.io/latest/docs/reference/config/security/jwt/#JWTRule).