feat: link istio authn to addressservice
Adds a custom IstioAuthentication class which just reads the user
context passed in a header by Istio and gets or creates a user from it.

Also adds a relation from Address to User. To make the API usable in a
demo in case of a non-Istio deployment, the relation to user is nullable
effectively providing an unauthenticated namespace of addresses.